+++ package.json
@@ -82,6 +82,7 @@
     "@types/webpack-config-utils": "^2.3.0",
     "@types/webpack-dev-server": "^3.1.1",
     "@types/webpack-manifest-plugin": "^1.3.2",
+    "@types/webpack-merge": "^4.1.3",
     "@types/webpack-node-externals": "^1.6.3",
     "autoprefixer": "9.1.5",
     "babel-core": "^6.26.0",
@@ -134,6 +135,7 @@
     "webpack-config-utils": "2.3.1",
     "webpack-dev-server": "3.1.10",
     "webpack-manifest-plugin": "2.0.4",
+    "webpack-merge": "^4.1.4",
     "webpack-node-externals": "1.7.2"
   },
   "devEngines": {
