+++ scripts/app/configs/webpack/common.ts
@@ -1,0 +1,259 @@
+import autoprefixer from "autoprefixer";
+import CaseSensitivePathsPlugin from "case-sensitive-paths-webpack-plugin";
+import ForkTsCheckerWebpackPlugin from "fork-ts-checker-webpack-plugin";
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
+import ModuleScopePlugin from "react-dev-utils/ModuleScopePlugin";
+import WatchMissingNodeModulesPlugin from "react-dev-utils/WatchMissingNodeModulesPlugin";
+import TsconfigPathsPlugin from "tsconfig-paths-webpack-plugin";
+import UglifyJsPlugin from "uglifyjs-webpack-plugin";
+import webpack, {Configuration, Module, Node, Options, Plugin, Resolve, RuleSetRule} from "webpack";
+import {getIfUtils, removeEmpty} from "webpack-config-utils";
+import Application from "../../Application";
+import Env from "../env";
+
+/**
+ * Utility functions to help segment configuration based on environment
+ */
+const {ifProduction, ifDevelopment} = getIfUtils(Env.current);
+
+/**
+ * Post CSS fixes
+ */
+const postCssPlugin = () => [
+    require("postcss-flexbugs-fixes"),
+    autoprefixer({
+        browsers: [
+            ">1%",
+            "last 4 versions",
+            "Firefox ESR",
+            "not ie < 9", // React doesn"t support IE8 anyway
+        ],
+        flexbox: "no-2009",
+    }),
+];
+
+/**
+ * All other files that aren't caught by the other loaders will go through this one.
+ * @description "file" loader makes sure those assets get served by WebpackDevServer.
+ * When you `import` an asset, you get its (virtual) filename.
+ * In production, they would get copied to the `build` folder.
+ * This loader doesn"t use a "test" so it will catch all modules
+ * that fall through the other loaders.
+ */
+const catchAllRule = {
+    exclude: [/\.(js|jsx|mjs)$/, /\.(ts|tsx)$/, /\.html$/, /\.json$/],
+    loader: require.resolve("file-loader"),
+    options: {
+        name: "[name].[hash:8].[ext]",
+        outputPath: "static/media/",
+    },
+};
+
+/**
+ * Handle css/scss
+ */
+const scssRule: RuleSetRule = {
+    test: /\.(scss|sass|css)$/i,
+    use: removeEmpty([
+        ifProduction(MiniCssExtractPlugin.loader),
+        ifDevelopment({loader: "style-loader", options: {sourceMap: true}}),
+        {loader: "css-loader", options: {sourceMap: true}},
+        {
+            loader: "postcss-loader",
+            options: {
+                ident: "postcss",
+                plugins: postCssPlugin,
+                sourceMap: true,
+            },
+        },
+        {loader: "sass-loader", options: {sourceMap: true}},
+    ]),
+};
+
+/**
+ * Define rule for static assets
+ * @description "url" loader works like "file" loader except that it embeds assets
+ * smaller than specified limit in bytes as data URLs to avoid requests.
+ */
+const staticFileRule: RuleSetRule = {
+    loader: require.resolve("url-loader"),
+    options: {
+        limit: 10000,
+        name: "static/media/[name].[hash:8].[ext]",
+    },
+    test: /\.(bmp|png|jpg|gif|svg|eot|ttf|woff|woff2)$/,
+};
+
+/**
+ * Define rule for transpiling TypeScript
+ * @description Disable type checker - we will use it in ForkTsCheckerWebpackPlugin
+ */
+const typeScriptRule: RuleSetRule = {
+    include: Application.srcPath,
+    test: /\.(ts|tsx)$/,
+    use: [
+        {
+            loader: require.resolve("ts-loader"),
+            options: {
+                configFile: Application.tsConfig,
+                transpileOnly: true,
+            },
+        },
+    ],
+};
+
+/**
+ * Define how source files are handled
+ */
+const module: Module = {
+    rules: [{
+        oneOf: [staticFileRule, typeScriptRule, scssRule, catchAllRule],
+    }],
+    strictExportPresence: true,
+};
+
+/**
+ * @description Some libraries import Node modules but don"t use them in the browser.
+ * Tell Webpack to provide empty mocks for them so importing them works.
+ */
+const node: Node = {
+    child_process: "empty",
+    dgram: "empty",
+    fs: "empty",
+    net: "empty",
+    tls: "empty",
+};
+
+/**
+ * Define build optimization options
+ */
+const optimization: Options.Optimization = {
+    minimize: ifProduction(),
+    minimizer: ifProduction([
+        /**
+         * Minify the code.
+         * Note: this won't work without ExtractTextPlugin.extract(..) in `loaders`.
+         *
+         * @env production
+         */
+        new UglifyJsPlugin({
+            cache: true,
+            parallel: true,
+            sourceMap: true, // TODO should we do this in prod??
+            uglifyOptions: {
+                compress: {
+                    comparisons: false,
+                    warnings: false,
+                },
+                output: {
+                    ascii_only: true,
+                    comments: false,
+                },
+            },
+        }),
+    ], []),
+};
+
+/**
+ * Define build performance options
+ */
+const performance: Options.Performance = {
+    hints: ifDevelopment("warning", false),
+};
+
+/**
+ * @description Plugins need to webpack to perform build
+ */
+const plugins: Plugin[] = removeEmpty([
+    /**
+     * This is necessary to emit hot updates (currently CSS only):
+     *
+     * @env development
+     */
+    ifDevelopment(new webpack.HotModuleReplacementPlugin(), undefined),
+    /**
+     * Watcher doesn"t work well if you mistype casing in a path so we use
+     * a plugin that prints an error when you attempt to do this.
+     * See https://github.com/facebookincubator/create-react-app/issues/240
+     *
+     * @env development
+     */
+    ifDevelopment(new CaseSensitivePathsPlugin(), undefined),
+
+    /**
+     * If you require a missing module and then `npm install` it, you still have
+     * to restart the development server for Webpack to discover it. This plugin
+     * makes the discovery automatic so you don"t have to restart.
+     * See https://github.com/facebookincubator/create-react-app/issues/186
+     *
+     * @env development
+     */
+    ifDevelopment(new WatchMissingNodeModulesPlugin(Application.nodeModulesPath), undefined),
+    /**
+     * Moment.js is an extremely popular library that bundles large locale files
+     * by default due to how Webpack interprets its code. This is a practical
+     * solution that requires the user to opt into importing specific locales.
+     * @url https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
+     * @env all
+     */
+    new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+    /**
+     * Perform type checking and linting in a separate process to speed up compilation
+     *
+     * @env all
+     */
+    ifProduction(new ForkTsCheckerWebpackPlugin({
+        async: false,
+        tsconfig: Application.tsConfig,
+        tslint: Application.tsLint,
+    }), new ForkTsCheckerWebpackPlugin({
+        async: false,
+        tsconfig: Application.tsConfig,
+        tslint: Application.tsLint,
+        watch: Application.srcPath,
+    })),
+]);
+
+/**
+ * @description Tell webpack how to resolve files and modules
+ * Prevents users from importing files from outside of src/ (or node_modules/).
+ * This often causes confusion because we only process files within src/ with babel.
+ * To fix this, we prevent you from importing files out of src/ -- if you'd like to,
+ * please link the files into your node_modules/ and let module-resolution kick in.
+ * Make sure your source files are compiled, as they will not be processed in any way.
+ */
+const resolve: Resolve = {
+    extensions: [
+        ".mjs",
+        ".web.ts",
+        ".ts",
+        ".web.tsx",
+        ".tsx",
+        ".web.js",
+        ".js",
+        ".json",
+        ".web.jsx",
+        ".jsx",
+    ],
+    modules: ["node_modules", Application.nodeModulesPath],
+    plugins: [
+        new ModuleScopePlugin(Application.srcPath, [Application.packagePath]),
+        new TsconfigPathsPlugin({ configFile: Application.tsConfig }),
+    ],
+};
+
+/**
+ * Common configuration
+ */
+const config: Configuration = {
+    bail: ifProduction(),
+    mode: ifProduction("production", "development"),
+    module,
+    node,
+    optimization,
+    performance,
+    plugins,
+    resolve,
+};
+
+export default config;
